change AbstractFileTextDataSourceComponent complete impl.
