+++ src/main/java/org/tuberlin/de/common/base/BaseDataSourceComponentText.java
@@ -2,6 +2,7 @@ package org.tuberlin.de.common.base;
 
 import org.tuberlin.de.common.model.Constants;
 import org.tuberlin.de.common.model.abstracts.datasource.file.AbstractFileDataSourceComponent;
+import org.tuberlin.de.common.model.abstracts.datasource.file.AbstractFileTextDataSourceComponent;
 import org.tuberlin.de.common.model.interfaces.JobGraph;
 import org.tuberlin.de.common.model.interfaces.datasource.DataSourceComponent;
 import org.tuberlin.de.common.model.interfaces.datasource.FileDataSourceComponent;
@@ -12,7 +13,7 @@ import java.util.Map;
 /**
  * Created by Malcolm-X on 26.12.2015.
  */
-public class BaseDataSourceComponentText extends AbstractFileDataSourceComponent implements DataSourceComponent, FileDataSourceComponent {
+public class BaseDataSourceComponentText extends AbstractFileTextDataSourceComponent implements DataSourceComponent, FileDataSourceComponent {
 
     public BaseDataSourceComponentText(JobGraph jobGraph, Map<String, Object> parameters){
         super();
@@ -20,19 +21,9 @@ public class BaseDataSourceComponentText extends AbstractFileDataSourceComponent
     }
 
     @Override
-    public void init(JobGraph jobGraph, Map<String, Object> parameters) {
-        super.init(jobGraph, parameters);
+    public String getJobSource() throws IllegalStateException {
+        return null;
     }
-
-    @Override
-    public String getJobSource() {
-        if (!initialized) throw new IllegalStateException("Must be initialized");
-        //TODO: Integrity checks
-        String result = jobGraph.getEnvironmentIdentifier() + ".readTextFile(" + parameters.get(Constants.COMPONENT_PATH_JSON) + ")";
-        result += ";";
-        return result;
-    }
-
 
 
     @Override
@@ -41,35 +32,4 @@ public class BaseDataSourceComponentText extends AbstractFileDataSourceComponent
         return super.verify();
     }
 
-    @Override
-    public Collection<String> getParents() throws IllegalStateException {
-        return null;
-    }
-
-    @Override
-    public Collection<String> getChildren() throws IllegalStateException {
-        return null;
-    }
-
-    @Override
-    public String getComponentKey() {
-        return null;
-    }
-
-    @Override
-    public String getInputType() {
-        return null;
-    }
-
-    @Override
-    public String getOutputType() {
-        return null;
-    }
-
-
-    @Override
-    public String getFilePath() {
-        //TODO integrity checks
-        return (String) parameters.get(FileDataSourceComponent.FILE_PATH_JSON);
-    }
 }
