+++ src/main/java/org/tuberlin/de/common/model/Constants.java
@@ -4,11 +4,20 @@ package org.tuberlin.de.common.model;
  * Created by Malcolm-X on 10.12.2015.
  */
 public class Constants {
+    /**
+     * which
+     */
     public static final String COMPONENT_TYPE_JSON = "COMPONENT_TYPE";
     public static final String TRANSFORMATION_TYPE_JSON = "TRANSFORMATION_TYPE";
     public static final String DATASOURCE_TYPE = "DATASOURCE_TYPE";
     public static final String DATASINK_TYPE = "DATASINK_TYPE";
+    /**
+     * If we implement the imports in a extra class not required anymore. In private class required
+     */
     public static final String COMPONENT_IMPORTS_JSON = "COMPONENT_IMPORTS_JSON";
+    /**
+     *
+     */
     public static final String COMPONENT_JOB_SOURCE_JSON = "COMPONENT_JOB_SOURCE_JSON";
     public static final String COMPONENT_SOURCE_JSON = "COMPONENT_SOURCE_JSON";
     public static final String COMPONENT_CHILDREN = "COMPONENT_CHILDREN_JSON";
